add documentation, NodeConnectionParameters.TemplateBehaviors
